combine cluster reg and es secrets in platform-operator
