+++ platform-operator/controllers/clusters/sync_prometheus.go
@@ -6,6 +6,8 @@ package clusters
 import (
 	"context"
 	"fmt"
+	"strings"
+
 	"github.com/Jeffail/gabs/v2"
 	clustersv1alpha1 "github.com/verrazzano/verrazzano/platform-operator/apis/clusters/v1alpha1"
 	"github.com/verrazzano/verrazzano/platform-operator/constants"
@@ -14,7 +16,6 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 	"sigs.k8s.io/yaml"
-	"strings"
 )
 
 const (
