+++ platform-operator/controllers/clusters/sync_registration_secret.go
@@ -5,10 +5,17 @@ package clusters
 
 import (
 	"context"
+	"fmt"
+
 	clusterapi "github.com/verrazzano/verrazzano/platform-operator/apis/clusters/v1alpha1"
+	"github.com/verrazzano/verrazzano/platform-operator/constants"
 	corev1 "k8s.io/api/core/v1"
+	k8net "k8s.io/api/networking/v1beta1"
+	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
+
+const vmiIngest = "vmi-system-es-ingest"
 
 // Create a registration secret with the managed cluster information.  This secret will
 // be used on the managed cluster to get information about itself, like the cluster name
@@ -41,8 +48,28 @@ func (r *VerrazzanoManagedClusterReconciler) createOrUpdateRegistrationSecret(vm
 // Mutate the secret, setting the kubeconfig data
 func (r *VerrazzanoManagedClusterReconciler) mutateRegistrationSecret(secret *corev1.Secret, manageClusterName string) error {
 	secret.Type = corev1.SecretTypeOpaque
+
+	// Get the info needed to build the elasicsearch secret
+	url, err := r.getElasticsearchURL()
+	if err != nil {
+		return err
+	}
+	vzSecret, err := r.getVzSecret()
+	if err != nil {
+		return err
+	}
+	tlsSecret, err := r.getTLSSecret()
+	if err != nil {
+		return err
+	}
+
+	// Build the secret data
 	secret.Data = map[string][]byte{
 		ManagedClusterNameKey: []byte(manageClusterName),
+		URLKey:                []byte(url),
+		CaBundleKey:           tlsSecret.Data[CaCrtKey],
+		UsernameKey:           vzSecret.Data[UsernameKey],
+		PasswordKey:           vzSecret.Data[PasswordKey],
 	}
 	return nil
 }
@@ -51,4 +78,50 @@ func (r *VerrazzanoManagedClusterReconciler) mutateRegistrationSecret(secret *co
 func GetRegistrationSecretName(vmcName string) string {
 	const registrationSecretSuffix = "-registration"
 	return generateManagedResourceName(vmcName) + registrationSecretSuffix
+}
+
+// Get the Elasticsearch URL.
+func (r *VerrazzanoManagedClusterReconciler) getElasticsearchURL() (URL string, err error) {
+	var Ingress k8net.Ingress
+	nsn := types.NamespacedName{
+		Namespace: constants.VerrazzanoSystemNamespace,
+		Name:      vmiIngest,
+	}
+	if err := r.Get(context.TODO(), nsn, &Ingress); err != nil {
+		return "", fmt.Errorf("Failed to fetch the VMI ingress %s/%s, %v", nsn.Namespace, nsn.Name, err)
+	}
+	if len(Ingress.Spec.Rules) == 0 {
+		return "", fmt.Errorf("VMI ingress %s/%s missing host entry in rule", nsn.Namespace, nsn.Name)
+	}
+	host := Ingress.Spec.Rules[0].Host
+	if len(Ingress.Spec.Rules) == 0 {
+		return "", fmt.Errorf("VMI ingress %s/%s host field is empty", nsn.Namespace, nsn.Name)
+	}
+	return fmt.Sprintf("https://%s:443", host), nil
+}
+
+// Get the Verrazzano secret
+func (r *VerrazzanoManagedClusterReconciler) getVzSecret() (corev1.Secret, error) {
+	var secret corev1.Secret
+	nsn := types.NamespacedName{
+		Namespace: constants.VerrazzanoSystemNamespace,
+		Name:      constants.Verrazzano,
+	}
+	if err := r.Get(context.TODO(), nsn, &secret); err != nil {
+		return corev1.Secret{}, fmt.Errorf("Failed to fetch the secret %s/%s, %v", nsn.Namespace, nsn.Name, err)
+	}
+	return secret, nil
+}
+
+// Get the system-tls secret
+func (r *VerrazzanoManagedClusterReconciler) getTLSSecret() (corev1.Secret, error) {
+	var secret corev1.Secret
+	nsn := types.NamespacedName{
+		Namespace: constants.VerrazzanoSystemNamespace,
+		Name:      constants.SystemTLS,
+	}
+	if err := r.Get(context.TODO(), nsn, &secret); err != nil {
+		return corev1.Secret{}, fmt.Errorf("Failed to fetch the secret %s/%s, %v", nsn.Namespace, nsn.Name, err)
+	}
+	return secret, nil
 }
