+++ platform-operator/test/integ/integ_test.go
@@ -169,13 +169,6 @@ var _ = ginkgo.Describe("Testing VMC creation and auto secret generation", func(
 		gomega.Eventually(secretExists, "30s", "5s").Should(gomega.BeTrue(),
 			fmt.Sprintf("The registration Secret %s should exist in %s", vzclusters.GetRegistrationSecretName(managedClusterName), vzMcNamespace))
 	})
-	ginkgo.It("Elasticsearch secret exists ", func() {
-		secretExists := func() bool {
-			return K8sClient.DoesSecretExist(vzclusters.GetElasticsearchSecretName(managedClusterName), vzMcNamespace)
-		}
-		gomega.Eventually(secretExists, "30s", "5s").Should(gomega.BeTrue(),
-			fmt.Sprintf("The Elasticsearch Secret %s should exist in %s", vzclusters.GetElasticsearchSecretName(managedClusterName), vzMcNamespace))
-	})
 	ginkgo.It("Agent secret exists ", func() {
 		secretExists := func() bool {
 			return K8sClient.DoesSecretExist(vzclusters.GetAgentSecretName(managedClusterName), vzMcNamespace)
